chore: Moved hooks outside the /lib/hooks directory, and within the /hooks folder.
