+++ app/components/search-filters/LocationPicker.jsx
@@ -6,8 +6,8 @@ import { Card, CardContent } from "@/components/ui/card";
 import { MapPin, Home, X } from "lucide-react";
 import { observable } from "@legendapp/state";
 import { observer } from "@legendapp/state/react";
-import { useListKeyboardNavigation } from "@/lib/hooks/useKeyboardNavigation";
-import useRestoreSelectionFromURLParam from "@/lib/hooks/useRestoreSelectionFromURLParam";
+import { useListKeyboardNavigation } from "@/hooks/useKeyboardNavigation";
+import useRestoreSelectionFromURLParam from "@/hooks/useRestoreSelectionFromURLParam";
 import { getLocationById } from "@/app/hotels/queryFunctions";
 import { useRouter, useSearchParams } from "next/navigation";
 import { cn } from "@/lib/utils";
